envio de cambios

+++ app/Http/Controllers/UsuariosController.php
@@ -18,7 +18,7 @@ class UsuariosController extends Controller
 
     public function index(UsuariosDataTable $dataTable)
     {
-        $roles=Role::all();
+        $roles=Role::where('name','!=','DECE')->get();
         $data = array('roles' => $roles );
         return $dataTable->render('usuarios.index',$data);
     }
@@ -26,12 +26,13 @@ class UsuariosController extends Controller
     public function guardar(Request $request)
     {
         $letras='/^[\pL\s\-]+$/u';
+        $rolDece=Role::where('name','DECE')->first();
         $request->validate([
             'name' => 'required|max:191|regex:'.$letras,
             'email' => 'required|string|email|max:191|unique:users',
             'password' => 'required|string|min:8|confirmed',
             "roles"    => "nullable|array",
-            "roles.*"  => "nullable|exists:roles,id",
+            "roles.*"  => "nullable|exists:roles,id|not_in:".$rolDece->id,
         ]);
         $user= User::create([
             'name' => $request->name,
@@ -48,7 +49,7 @@ class UsuariosController extends Controller
     public function editar($idUser)
     {
         $user=User::findOrFail($idUser);
-        $roles=Role::all();
+        $roles=Role::where('name','!=','DECE')->get();
         return view('usuarios.editar',['user'=>$user,'roles'=>$roles]);
     }
     public function eliminar(Request $request)
@@ -56,10 +57,10 @@ class UsuariosController extends Controller
         $request->validate([
             'user' => 'required|exists:users,id',
         ]);
-        
+        $user=User::findOrFail($request->user);
+        $this->authorize('actualizar', $user);
         try {
             DB::beginTransaction();
-            $user=User::findOrFail($request->user);
             if(Auth::user()->id!=$user->id){
                 $user->delete();
                 DB::commit();
@@ -78,16 +79,18 @@ class UsuariosController extends Controller
     public function actualizar(Request $request)
     {
         $letras='/^[\pL\s\-]+$/u';
+        $rolDece=Role::where('name','DECE')->first();
         $request->validate([
             'usuario'=>'required|exists:users,id',
             'name' => 'required|max:255|regex:'.$letras,
             'email' => 'required|string|email|max:255|unique:users,email,'.$request->usuario,
             'password' => 'nullable|string|min:8|confirmed',
             "roles"    => "nullable|array",
-            "roles.*"  => "nullable|exists:roles,id",
+            "roles.*"  => "nullable|exists:roles,id|not_in:".$rolDece->id,
         ]);
-
+            
         $user=User::findOrFail($request->usuario);
+        $this->authorize('actualizar', $user);
         $user->name=$request->name;
         $user->email=$request->email;
         if($request->password){
